[chore] api / router 모듈 분리
