+++ week02/level2/src/api/auth.ts
@@ -1,0 +1,10 @@
+import { Request, Response } from "express";
+
+const signUp = (req: Request, res: Response) => {
+  return res.status(201).json({
+    status: 201,
+    message: "회원가입 성공",
+  });
+}
+
+export default signUp;
