--- week02/level2/src/api/like.ts
@@ -1,12 +1,0 @@
-import express, { Request, Response, Router } from "express";
-
-const router: Router = express.Router();
-
-router.get("/", (req: Request, res: Response) => {
-  return res.status(200).json({
-    status: 200,
-    message: "좋아요 조회 성공",
-  });
-});
-
-module.exports = router;
