+++ week02/level2/src/router/authRouter.ts
@@ -1,0 +1,8 @@
+import express, { Router } from 'express';
+import signUp from '../api/auth';
+
+const router: Router = express.Router();
+
+router.get('/', signUp);
+
+export default router;
