+++ week02/level2/src/router/blogRouter.ts
@@ -1,0 +1,9 @@
+import express, { Router } from 'express';
+import { selectPost, likePost } from '../api/blog';
+
+const router: Router = express.Router();
+
+router.get('/', selectPost);
+router.get('/like', likePost);
+
+export default router;
