+++ week02/level2/src/router/userRouter.ts
@@ -1,0 +1,8 @@
+import express, { Router } from 'express';
+import selectUser from '../api/user';
+
+const router: Router = express.Router();
+
+router.get('/', selectUser);
+
+export default router;
